:art: Restrukturert preview-provider
